Move utilities to utils/fs
